services/chat_service/requirements.txt
-r ../../requirements-common.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
httpx==0.24.1
websockets==11.0.3
